requirements.txt
python-dotenv
django-allauth
